+++ server/text_generation_server/models/vlm_causal_lm.py
@@ -1,11 +1,7 @@
-import os
 import torch
-import tempfile
-import requests
 from PIL import Image
 from io import BytesIO
 
-from contextlib import contextmanager
 
 from opentelemetry import trace
 from typing import Iterable, Optional, Tuple, List, Type, Dict
@@ -196,8 +192,6 @@ class VlmCausalLMBatch(FlashCausalLMBatch):
                         images.append([image])
                 elif chunk_type == "video":
                     if config.model_type == "qwen2_vl":
-                        # For now, treat video URLs as special tokens
-                        # This will be processed in the text replacement section below
                         pass
                 else:
                     raise RuntimeError(f"Invalid chunk type {chunk_type}")
@@ -222,13 +216,11 @@ class VlmCausalLMBatch(FlashCausalLMBatch):
                     )
                     image_id += 1
                 elif chunk_type == "video" and config.model_type == "qwen2_vl":
-                    # Download and process video in a temporary context
-                    with cls.temp_video_download(chunk.video) as local_path:
-                        # Now the video is available at local_path for processing
-                        full_text += f"<video>{local_path}</video>"
+                    from text_generation_server.models.custom_modeling.qwen2_vl import process_qwen_video
+                    text, _ = process_qwen_video(chunk.video)
+                    full_text += text
 
             full_text = image_text_replacement_fixup(config, full_text)
-
             batch_inputs.append(full_text)
             max_truncation = max(max_truncation, r.truncate)
 
@@ -277,27 +269,6 @@ class VlmCausalLMBatch(FlashCausalLMBatch):
             batch.image_sizes = None
             batch.image_grid_thw = None
         return batch
-    
-    @staticmethod
-    @contextmanager
-    def temp_video_download(url: str) -> str:
-        """Downloads video to a temporary file and cleans it up after use."""
-        with tempfile.NamedTemporaryFile(suffix=os.path.splitext(url)[1], delete=False) as tmp_file:
-            try:
-                # Download video
-                with requests.get(url, stream=True) as r:
-                    r.raise_for_status()
-                    for chunk in r.iter_content(chunk_size=8192):
-                        if chunk:
-                            tmp_file.write(chunk)
-                tmp_file.flush()
-                yield tmp_file.name
-            finally:
-                # Clean up temp file
-                try:
-                    os.unlink(tmp_file.name)
-                except OSError:
-                    pass
 
 class VlmCausalLM(FlashCausalLM):
     def __init__(
